+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master
 
 - Changed the default value of `-[UIApplication appChannel]` to `"AppStore"`, before is `"App Store"`.
+- Replace whitespace to `"-"` for the default value of `-[UIApplication appName]`.
 - Renamed NSURLComponents `-queryItemsDictionary` to `-queryParameters`, `-addQueryItemsDictionary:` to `-addQueryParameters:`.
 - Renamed NSURL `-queryDictionary` to `-queryParameters`, `-URLByAddingQueryDictionary:` to `-URLByAddingQueryParameters:`.
 - Renamed NSString `-queryDictionary` to `-URLQueryParameters`, `-stringByAddingQueryDictionary:` to `-stringByAddingURLQueryParameters:`